migrated running dose gen to dose gen utils since the plan export is very specific to the dose gen algorithm
