+++ brachyutils/planning/plan_utils.py
@@ -1810,60 +1810,6 @@ config do not match for structure {struc.name}"
                 continue
             structure.optimization_config = None
 
-    def run_dose_generation(
-        self,
-        dir_export: str | Path = None,
-        dose_generator_obj: BrachyDoseGenerator = None,
-        generate_dose_rate_maps: bool = False,
-        export_config_brachyplan: ExportConfig_BrachyPlan = None,
-        ):
-        r"""
-        ### Purpose:
-        - to run the dose generation for the plan and update the combined dose.
-        ### Inputs:
-        - dose_generator_obj := the BrachyDoseGenerator object to be used for dose generation
-        - generate_dose_rate_maps := whether to generate dose rate maps for each dwell position.
-        If True, the dose_rate_dict will be populated with the dose rate maps for each dwell position.
-        - dir_export := the directory used for exporting the dosimetry setup and the generated dose maps.
-        if None, "temp_data/tg43/"
-        """
-        plan_name = self.phantom.pth_image.stem
-        if "." in plan_name:
-            plan_name = plan_name.split(".")[0]
-
-        if dir_export is None:
-            # XXX check the full path with resolve to be correct.
-            dir_export = Path("temp_data/tg43/")/plan_name
-
-        dir_export = Path(dir_export)
-        if dose_generator_obj is None:
-            dose_generator_obj = DoseTG43(dir_plan_export=dir_export)
-
-        if not isinstance(dose_generator_obj, BrachyDoseGenerator):
-            raise ValueError("dose_generator_obj should be an instance of BrachyDoseGenerator")
-        # export the plan for dosimetry
-        if export_config_brachyplan is None:
-            export_config_brachyplan = ExportConfig_BrachyPlan(
-                dir_export=dir_export,
-                export_config_egsphant=True,
-                export_config_planfile=True,
-                export_config_macfile=True,
-            )
-        self.export_brachy_plan(export_config_brachyplan)
-        # call the dose generator to generate the dose maps
-        dose_generator_obj.generate_dose(
-            output_dose_per_dwell= "dose_rate" if generate_dose_rate_maps else False,
-        )
-        # load the generated dose maps and update the plan
-        if generate_dose_rate_maps:
-            self.load_dose_rate_dict(
-                dir_dose_rate=dir_export,
-            )
-        else:
-            self.combined_dose = BrachyDose(
-                export_config_brachyplan.export_config_macfile.pth_combined.with_suffix(".seq.nrrd")
-                )            
-
     def get_dose_rate_matrices_for_catheter(
         self,
         catheter_index: int
